add resources data and clean up seedData
